-Added toasts, used only for achievements for now -Added Achievements screen -Added Stats screen -Added recipes for ore blocks -Now, the original lang assets will be loaded, and they will be needed from now on -Now, the gui tiles lighting will be accurate to the original -Now, the Select World screen will be accurate to the original -Fixed player and mob kills stats not being obtainable -Fixed Change Profile screen not changing the stats counter used username -Fixed achievements map being parsed incorrectly -Fixed build sword achievement not being obtainable

+++ source/stats/StatFormatter.cpp
@@ -39,7 +39,7 @@ DecimalFormatter::DecimalFormatter(int precision) : m_precision(precision)
 {
 }
 
-std::string DecimalFormatter::format(int value) const
+std::string DecimalFormatter::format(double value) const
 {
 	std::stringstream ss;
 	ss.imbue(std::locale(""));
